+++ src/api/controller.ts
@@ -1,0 +1,7 @@
+import { Router } from 'express';
+
+export abstract class Controller {
+  constructor(public readonly route: string) {}
+
+  public abstract getRouter(): Router;
+}