+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-#  (2020-08-26)
+#  (2020-08-27)
 
 
 ### Features
@@ -6,15 +6,17 @@
 - CORS middleware ([8372ca3](https://github.com/Miczeq22/brainflash-api/commit/8372ca37caebd56533d39955684fd793e5582dbf))
 - Domain tools ([1ed8e4c](https://github.com/Miczeq22/brainflash-api/commit/1ed8e4cb026d418a42e511f144e77165af732d08))
 - Error handler middleware ([d62e86f](https://github.com/Miczeq22/brainflash-api/commit/d62e86f46c1b8422a5b32d85680b8e12d2383a98))
+- Register User Command ([cf520a7](https://github.com/Miczeq22/brainflash-api/commit/cf520a7001d89b3595bc08abad3e4ecb6b392353))
 - Security Middleware ([08ea082](https://github.com/Miczeq22/brainflash-api/commit/08ea0826dfcd30484a390e23665a6e443a624dd0))
 - Server bootstrap ([9a1e2f2](https://github.com/Miczeq22/brainflash-api/commit/9a1e2f21d6db5e1b62a5df019e4a8bfb975fb921))
 - User Registration Domain ([afd329a](https://github.com/Miczeq22/brainflash-api/commit/afd329aa493813c5f85492ec6582898a363b2ac4))
+- User registration endpoint ([4f08623](https://github.com/Miczeq22/brainflash-api/commit/4f086234f4aad26f6b1b3fd6f84b16f2a5ef2278))
 
 
 ### Tooling
 
 - Added postgres database layer ([164a310](https://github.com/Miczeq22/brainflash-api/commit/164a31087b591c996669ca37681223cd52067272))
-- Processing tools ([fcd01ac](https://github.com/Miczeq22/brainflash-api/commit/fcd01acb72fb3b6d5d5f2a214ce88490af740703))
+- Processing tools ([4b91c8d](https://github.com/Miczeq22/brainflash-api/commit/4b91c8d4fbd237d3d4303fddc17155a8b262119f))
 
 
 ### Documentation